debian/patches: Don't bother with PowerPC compilation of mpeg2dec.
We are not building `mpeg2dec` and, if necessary, the other parts of the
patch can be appropriately forwarded to the packager or upstream.

Note that it may be possible to work around the necessity of mpeg2dec by
passing the option `--enable-ff-mpeg2` to the main configure script.
I have not yet investigated this possibility.