Force handbrake to link against the system's libav.
Note that here we use a very dirty hack, namely, taking the file
`audioconvert.h` from libav's git tree (which I was told is not a public
interface), and putting it `libhb` directory.

We also pass the `/usr/include/libavcodec` directory as an include dir for
the other headers of libav to be found.