Use libav/ffmpeg for reading metadata
This substitutes mp4v2 for reading metadata from files. It is currently full
of FIXME's and it was only lightly tested with a pair of MP4 files.

It needs a lot of love, but it is my first try at learning how to use
libav/ffmpeg's API.

Note to self: in a call to `avformat_open_input`, the `AVFormatContext`
pointer that is passed must be initialized to `NULL` or bad things happen
(e.g., segfaults).

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>